requirements.txt
langchain==0.2.11
langchain-community==0.2.11
langchain-text-splitters==0.2.2
langchain-pinecone==0.1.3
langchain-huggingface==0.0.7
pinecone-client==3.2.2
python-dotenv==1.0.1
fastapi==0.111.0
pydantic==2.7.1
PyPDF2==3.0.1
huggingface-hub==0.23.4
transformers==4.41.1
torch>=2.0.0
